Refactored Some files to minimize clutter
