+++ bin/.gitignore
@@ -1,4 +1,2 @@
-/InitClass.class
-/Tetromino.class
+/Pieces/
 /finalProject/
-/sedgewick/
